integration_tests/test-requirements.txt
docker-compose
kombu
openapi-spec-validator
psycopg2-binary  # from sqlalchemy
pyhamcrest
pytest
requests
sqlalchemy
https://github.com/wazo-platform/wazo-lib-rest-client/archive/master.zip
https://github.com/wazo-platform/wazo-agentd-client/archive/master.zip
https://github.com/wazo-platform/xivo-dao/archive/master.zip
https://github.com/wazo-platform/xivo-lib-python/archive/master.zip
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
